+++ JackpotMall/src/main/webapp/WEB-INF/jsp/mall/Main.jsp
@@ -4,34 +4,74 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
-    <meta charset="UTF-8">
-    <meta charset="utf-8">
-    <title>JackpotMall</title>
-   
-    
+<meta charset="UTF-8">
+<meta charset="utf-8">
+<title>JackpotMall</title>
+
+
 </head>
 <body>
- <jsp:include page="../include/header.jsp" />
+	<jsp:include page="../include/header.jsp" />
 
 
 
-    <section>
-      <div class="product-wrapper">
-          <div class="pro_photos"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-          <div class="pro_photos"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-          <div class="pro_photos"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-          <div class="pro_photos"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-          <div class="pro_photos"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-      </div>
-      <div class="product-wrapper2">
-        <div class="pro_photos2"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-        <div class="pro_photos2"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-        <div class="pro_photos2"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-        <div class="pro_photos2"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-        <div class="pro_photos2"><img src="images/shopping/unnamed.jpg"width=110px height="140px"><p>상품이름</p> <p>가격</p></div>
-    </div>
-    </section>
+	<section>
+	<div class="product-wrapper">
+		<div class="pro_photos">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+	</div>
+	<div class="product-wrapper2">
+		<div class="pro_photos2">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos2">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos2">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos2">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+		<div class="pro_photos2">
+			<img src="images/shopping/unnamed.jpg" width=110px height="140px">
+			<p>상품이름</p>
+			<p>가격</p>
+		</div>
+	</div>
+	</section>
 
-<jsp:include page="../include/footer.jsp" />
+	<jsp:include page="../include/footer.jsp" />
 </body>
 </html>
