+++ JackpotMall/src/main/webapp/css/Login/login.css
@@ -1,6 +1,7 @@
 @charset "UTF-8";
 
 
+
         .login {
             width: 300px;
             height: 250px;
